Add automated deployment via Travis CI

+++ build_runtime.sh
@@ -1,7 +1,17 @@
 #!/bin/bash
-VERSION=${VERSION:=3.5.1}
-rm -r R/
-unzip R-$VERSION.zip -d R/
+
+set -euo pipefail
+
+VERSION=$1
+
+if [ -z "$VERSION" ];
+then
+    echo 'version number required'
+    exit 1
+fi
+
+rm -rf R/
+unzip -q R-$VERSION.zip -d R/
 rm -r R/doc/manual/
 #remove some libraries to save space
 recommended=(boot class cluster codetools foreign KernSmooth lattice MASS Matrix mgcv nlme nnet rpart spatial survival)
@@ -10,6 +20,5 @@ do
    rm -r R/library/$package/
 done
 chmod -R 755 bootstrap runtime.r R/
-rm runtime.zip
-zip -r runtime.zip runtime.r bootstrap R/
-aws lambda publish-layer-version --layer-name r-runtime --zip-file fileb://runtime.zip
+rm -f runtime.zip
+zip -r -q runtime.zip runtime.r bootstrap R/
